refactor to to add services/file.js
